requirements.txt
Django==4.2.11
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
PyJWT==2.8.0
python-decouple==3.8
python-jose==3.3.0
requests==2.32.3
asgiref==3.8.1
Authlib==1.3.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
ecdsa==0.19.0
idna==3.7
pyasn1==0.6.0
pycparser==2.22
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
